+++ libsxe/lib-sxe-pool/test/test-sxe-pool.c
@@ -19,6 +19,7 @@
  * THE SOFTWARE.
  */
 
+#include <inttypes.h>
 #include <string.h>
 #include <stdio.h>
 #include <stddef.h>
@@ -141,7 +142,7 @@ main(void)
     /* Initialization causes expected state
      */
     ok((size = sxe_pool_size(4, sizeof(*pool), TEST_STATE_NUMBER_OF_STATES)) >= 4 * sizeof(*pool),
-       "Expect pool size %u to be at least the size of the array %u", size, 4 * sizeof(*pool));
+       "Expect pool size %"PRIuPTR" to be at least the size of the array %"PRIuPTR"", (uintptr_t)size, (uintptr_t)4 * sizeof(*pool));
     SXEA10((base[0] = malloc(size)) != NULL,                                  "Couldn't allocate memory for 1st copy of pool");
     pool = sxe_pool_construct(base[0], "cesspool", 4, sizeof(*pool), TEST_STATE_NUMBER_OF_STATES, SXE_POOL_OPTION_TIMED);
     SXEA10((base[1] = malloc(size)) != NULL,                                  "Couldn't allocate memory for 2nd copy of pool");
